+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Changes in version 1.4.4 (October 6, 2022)
 
+- Fix `plot.designDiagram()` so that it aligns with `vctrs` usage in `ggraph` version 2.1.0.
 - Fix bug in *DD.R* so `inherits(.,"formula")` is used instead of `class(.)!="formula"`
 - Fix bug in *DD.R* on check of factors inside design matrices